+++ components/layouts/ChatManager.tsx
@@ -3,11 +3,7 @@ import React, { useEffect, useRef, useState } from 'react'
 import ChatContainer from 'components/chat/ChatContainer'
 import ChatNav from 'components/chat/ChatNav'
 import { compareJeuneChat, JeuneChat } from 'interfaces/jeune'
-import {
-  getChatCredentials,
-  observeConseillerChats,
-  signIn,
-} from 'services/messages.service'
+import { observeConseillerChats } from 'services/messages.service'
 import styles from 'styles/components/Layouts.module.css'
 import { useChatCredentials } from 'utils/chat/chatCredentialsContext'
 import { useConseiller } from 'utils/conseiller/conseillerContext'
@@ -18,13 +14,11 @@ const CHEMIN_DU_SON = '/sounds/notification.mp3'
 interface ChatManagerProps {
   displayChat: boolean
   setHasMessageNonLu: (value: boolean) => void
-  pageEstMessagerie?: boolean
 }
 
 export default function ChatManager({
   displayChat,
   setHasMessageNonLu,
-  pageEstMessagerie,
 }: ChatManagerProps) {
   const [conseiller] = useConseiller()
   const [portefeuille] = usePortefeuille()
@@ -102,7 +96,6 @@ export default function ChatManager({
         <ChatContainer
           jeunesChats={chats}
           menuState={[showChatNav, setShowChatNav]}
-          messagerieFullScreen={pageEstMessagerie}
         />
       </aside>
