+++ tests/components/ChatContainer.test.tsx
@@ -52,7 +52,13 @@ describe('<ChatContainer />', () => {
   describe('Messagerie', () => {
     it('affiche la messagerie', async () => {
       // When
-      renderWithContexts(<ChatContainer jeunesChats={jeunesChats} />, {})
+      renderWithContexts(
+        <ChatContainer
+          jeunesChats={jeunesChats}
+          menuState={[false, () => {}]}
+        />,
+        {}
+      )
 
       // Then
       expect(
@@ -68,9 +74,15 @@ describe('<ChatContainer />', () => {
     beforeEach(async () => {
       // When
       await act(async () => {
-        renderWithContexts(<ChatContainer jeunesChats={jeunesChats} />, {
-          customCurrentJeune: { id: jeunes[2].id },
-        })
+        renderWithContexts(
+          <ChatContainer
+            jeunesChats={jeunesChats}
+            menuState={[false, () => {}]}
+          />,
+          {
+            customCurrentJeune: { id: jeunes[2].id },
+          }
+        )
       })
     })
 
@@ -90,7 +102,10 @@ describe('<ChatContainer />', () => {
 
   describe('listes de diffusion', () => {
     beforeEach(async () => {
-      renderWithContexts(<ChatContainer jeunesChats={[]} />, {})
+      renderWithContexts(
+        <ChatContainer jeunesChats={[]} menuState={[false, () => {}]} />,
+        {}
+      )
     })
 
     it('permet d’accéder aux messages envoyés aux listes de diffusion', async () => {
@@ -127,8 +142,3 @@ describe('<ChatContainer />', () => {
     })
   })
 })
-
-const resizeWindow = (x: number) => {
-  window = Object.assign(window, { innerWidth: x })
-  window.dispatchEvent(new Event('resize'))
-}
