+++ tests/utils/chatsContext.test.tsx
@@ -4,11 +4,8 @@ import React from 'react'
 import { unConseiller } from 'fixtures/conseiller'
 import { desItemsJeunes, extractBaseJeune, unJeuneChat } from 'fixtures/jeune'
 import { JeuneChat } from 'interfaces/jeune'
-import {
-  getChatCredentials,
-  observeConseillerChats,
-  signIn,
-} from 'services/messages.service'
+import { observeConseillerChats } from 'services/messages.service'
+import { ChatCredentialsProvider } from 'utils/chat/chatCredentialsContext'
 import { ChatsProvider } from 'utils/chat/chatsContext'
 import { ConseillerProvider } from 'utils/conseiller/conseillerContext'
 import { PortefeuilleProvider } from 'utils/portefeuilleContext'
@@ -44,11 +41,6 @@ describe('ChatsProvider', () => {
 
   beforeEach(async () => {
     // When
-    ;(getChatCredentials as jest.Mock).mockResolvedValue({
-      token: 'tokenFirebase',
-      cleChiffrement: 'cleChiffrement',
-    })
-    ;(signIn as jest.Mock).mockResolvedValue({})
     ;(observeConseillerChats as jest.Mock).mockImplementation(
       (_jeune, _cle, fn) => {
         updateChatsRef = fn
@@ -61,6 +53,10 @@ describe('ChatsProvider', () => {
 
   describe('cas nominal', () => {
     const conseiller = unConseiller({ notificationsSonores: true })
+    const credentials = {
+      token: 'tokenFirebase',
+      cleChiffrement: 'cleChiffrement',
+    }
     beforeEach(async () => {
       // When
       await act(async () =>
@@ -69,24 +65,16 @@ describe('ChatsProvider', () => {
             <link rel='icon' href='/favicon.png' />
             <ConseillerProvider conseiller={conseiller}>
               <PortefeuilleProvider portefeuille={portefeuille}>
-                <ChatsProvider>
-                  <div />
-                </ChatsProvider>
+                <ChatCredentialsProvider credentials={credentials}>
+                  <ChatsProvider>
+                    <div />
+                  </ChatsProvider>
+                </ChatCredentialsProvider>
               </PortefeuilleProvider>
             </ConseillerProvider>
           </>
         )
       )
-    })
-
-    it('récupère les informations pour se connecter à la messagerie', async () => {
-      // Then
-      expect(getChatCredentials).toHaveBeenCalledWith()
-    })
-
-    it('se connecte à la messagerie', () => {
-      // Then
-      expect(signIn).toHaveBeenCalledWith('tokenFirebase')
     })
 
     it('observe les conversations', () => {
